+++ hydra/_internal/config_loader_impl.py
@@ -5,7 +5,6 @@ import re
 import sys
 import warnings
 from collections import defaultdict
-from dataclasses import dataclass
 from textwrap import dedent
 from typing import Any, Dict, List, MutableSequence, Optional
 
@@ -35,11 +34,7 @@ from hydra.errors import ConfigCompositionException, MissingConfigException
 from hydra.plugins.config_source import ConfigLoadError, ConfigResult, ConfigSource
 from hydra.types import RunMode
 
-
-@dataclass
-class SplitOverrides:
-    config_group_overrides: List[Override]
-    config_overrides: List[Override]
+from .deprecation_warning import deprecation_warning
 
 
 class ConfigLoaderImpl(ConfigLoader):
@@ -167,6 +162,11 @@ class ConfigLoaderImpl(ConfigLoader):
                 primary_config = loaded.config
         else:
             primary_config = OmegaConf.create()
+
+        if not OmegaConf.is_dict(primary_config):
+            raise ConfigCompositionException(
+                f"primary config '{config_name}' must be a DictConfig, got {type(primary_config).__name__}"
+            )
 
         def is_searchpath_override(v: Override) -> bool:
             return v.get_key_element() == "hydra.searchpath"
@@ -395,9 +395,9 @@ class ConfigLoaderImpl(ConfigLoader):
         ret = repo.load_config(config_path=config_path)
         assert ret is not None
 
-        if not isinstance(ret.config, DictConfig):
+        if not OmegaConf.is_config(ret.config):
             raise ValueError(
-                f"Config {config_path} must be a Dictionary, got {type(ret).__name__}"
+                f"Config {config_path} must be an OmegaConf config, got {type(ret.config).__name__}"
             )
 
         if not ret.is_schema_source:
@@ -425,7 +425,7 @@ class ConfigLoaderImpl(ConfigLoader):
                             )
                         )
                     else:
-                        warnings.warn(
+                        deprecation_warning(
                             dedent(
                                 f"""\
 
@@ -433,7 +433,6 @@ class ConfigLoaderImpl(ConfigLoader):
                                 This behavior is deprecated in Hydra 1.1 and will be removed in Hydra 1.2.
                                 See {url} for migration instructions."""
                             ),
-                            category=UserWarning,
                             stacklevel=11,
                         )
 
@@ -445,14 +444,16 @@ class ConfigLoaderImpl(ConfigLoader):
                         default.config_path is not None
                         and default.config_path.startswith("hydra/")
                     )
+                    config = ret.config
                     if (
                         default.primary
-                        and "hydra" in ret.config
+                        and isinstance(config, DictConfig)
+                        and "hydra" in config
                         and not hydra_config_group
                     ):
-                        hydra = ret.config.pop("hydra")
+                        hydra = config.pop("hydra")
 
-                    merged = OmegaConf.merge(schema.config, ret.config)
+                    merged = OmegaConf.merge(schema.config, config)
                     assert isinstance(merged, DictConfig)
 
                     if hydra is not None:
@@ -470,6 +471,7 @@ class ConfigLoaderImpl(ConfigLoader):
         if (
             not default.primary
             and config_path != "hydra/config"
+            and isinstance(res.config, DictConfig)
             and OmegaConf.select(res.config, "hydra.searchpath") is not None
         ):
             raise ConfigCompositionException(