+++ tests/defaults_list/test_defaults_tree.py
@@ -741,6 +741,53 @@ def test_two_group_defaults_different_pkgs(
             ),
             id="legacy_override_hydra+external",
         ),
+        param(
+            "legacy_override_hydra2",
+            [],
+            DefaultsTreeNode(
+                node=VirtualRoot(),
+                children=[
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="hydra/config"),
+                        children=[
+                            GroupDefault(group="help", value="custom1"),
+                            GroupDefault(group="output", value="disabled"),
+                            ConfigDefault(path="_self_"),
+                        ],
+                    ),
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="legacy_override_hydra2"),
+                        children=[ConfigDefault(path="_self_")],
+                    ),
+                ],
+            ),
+            id="legacy_override_hydra+external",
+        ),
+        param(
+            "legacy_override_hydra_wrong_order",
+            [],
+            DefaultsTreeNode(
+                node=VirtualRoot(),
+                children=[
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="hydra/config"),
+                        children=[
+                            GroupDefault(group="help", value="custom1"),
+                            GroupDefault(group="output", value="default"),
+                            ConfigDefault(path="_self_"),
+                        ],
+                    ),
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="legacy_override_hydra_wrong_order"),
+                        children=[
+                            GroupDefault(group="group1", value="file1"),
+                            ConfigDefault(path="_self_"),
+                        ],
+                    ),
+                ],
+            ),
+            id="legacy_override_hydra_wrong_order",
+        ),
     ],
 )
 def test_legacy_hydra_overrides_from_primary_config(
@@ -761,6 +808,52 @@ def test_legacy_hydra_overrides_from_primary_config(
             expected=expected,
             prepend_hydra=True,
         )
+
+
+@mark.parametrize(
+    "config_name,overrides,expected",
+    [
+        param(
+            "legacy_override_hydra2",
+            [],
+            DefaultsTreeNode(
+                node=VirtualRoot(),
+                children=[
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="hydra/config"),
+                        children=[
+                            GroupDefault(group="help", value="custom1"),
+                            GroupDefault(group="output", value="disabled"),
+                            ConfigDefault(path="_self_"),
+                        ],
+                    ),
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="legacy_override_hydra2"),
+                        children=[ConfigDefault(path="_self_")],
+                    ),
+                ],
+            ),
+            id="legacy_override_hydra+external",
+        ),
+    ],
+)
+def test_legacy_hydra_overrides_from_primary_config_2(
+    config_name: str, overrides: List[str], expected: DefaultsTreeNode, recwarn: Any
+) -> None:
+    """
+    Override two Hydra config groups using legacy notation
+    """
+
+    _test_defaults_tree_impl(
+        config_name=config_name,
+        input_overrides=overrides,
+        expected=expected,
+        prepend_hydra=True,
+    )
+
+    assert len(recwarn) == 2
+    assert "Invalid overriding of hydra/help:" in recwarn.list[0].message.args[0]
+    assert "Invalid overriding of hydra/output:" in recwarn.list[1].message.args[0]
 
 
 @mark.parametrize(
@@ -846,6 +939,52 @@ def test_group_default_with_appended_experiment(
         config_name=config_name,
         input_overrides=overrides,
         expected=expected,
+    )
+
+
+@mark.parametrize(
+    "config_name,overrides,expected",
+    [
+        param(
+            "override_hydra",
+            ["+experiment=override_config_group"],
+            DefaultsTreeNode(
+                node=VirtualRoot(),
+                children=[
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="hydra/config"),
+                        children=[
+                            GroupDefault(group="help", value="default"),
+                            GroupDefault(group="output", value="disabled"),
+                            ConfigDefault(path="_self_"),
+                        ],
+                    ),
+                    DefaultsTreeNode(
+                        node=ConfigDefault(path="override_hydra"),
+                        children=[
+                            GroupDefault(group="group1", value="file2"),
+                            ConfigDefault(path="_self_"),
+                            GroupDefault(
+                                group="experiment", value="override_config_group"
+                            ),
+                        ],
+                    ),
+                ],
+            ),
+            id="override_hydra_with_experiment",
+        ),
+    ],
+)
+def test_experiment_where_primary_config_has_override(
+    config_name: str,
+    overrides: List[str],
+    expected: DefaultsTreeNode,
+) -> None:
+    _test_defaults_tree_impl(
+        config_name=config_name,
+        input_overrides=overrides,
+        expected=expected,
+        prepend_hydra=True,
     )